Remove WorkflowManager
