+++ src/beamlime/services/data_reduction.py
@@ -12,7 +12,6 @@ from beamlime.core.message import CONFIG_STREAM_ID
 from beamlime.core.orchestrating_processor import OrchestratingProcessor
 from beamlime.handlers.config_handler import ConfigHandler
 from beamlime.handlers.data_reduction_handler import ReductionHandlerFactory
-from beamlime.handlers.workflow_manager import WorkflowManager
 from beamlime.kafka.routes import RoutingAdapterBuilder
 from beamlime.service_factory import DataServiceBuilder, DataServiceRunner
 
@@ -31,15 +30,12 @@ def make_reduction_service_builder(
     )
     _ = get_config(instrument)  # Load the module to register the instrument
     instrument_config = instrument_registry[instrument]
-    workflow_manager = WorkflowManager(
-        processor_factory=instrument_config.processor_factory,
-        source_to_key=instrument_config.source_to_key,
-    )
     service_name = 'data_reduction'
     config_handler = ConfigHandler(service_name=service_name)
     handler_factory = ReductionHandlerFactory(
-        workflow_manager=workflow_manager,
         f144_attribute_registry=instrument_config.f144_attribute_registry,
+        processor_factory=instrument_config.processor_factory,
+        source_to_key=instrument_config.source_to_key,
     )
     builder = DataServiceBuilder(
         instrument=instrument,
